// spooltracking/src/app/services/api-clients/model/spool.ts
/**
 * SpooltrackingAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SpoolBrand } from './spoolBrand';


export interface Spool { 
    id?: string;
    brandId?: string;
    brand?: SpoolBrand;
    material?: string | null;
    color?: string | null;
    weight?: number;
}